Coresight: Add coresight dummy driver
Some Coresight devices that kernel don't have permission to access or
configure. For these devices, a dummy driver is needed to register them as
Coresight devices. The module may also be used to define components that may
not have any programming interfaces (e.g, static links), so that paths can be
created in the driver. It provides Coresight API for operations on dummy
devices, such as enabling and disabling them. It also provides the Coresight
dummy sink/source paths for debugging.

Signed-off-by: Hao Zhang <quic_hazha@quicinc.com>