Coresight: Add coresight TGU driver
This driver provides support for Trigger Generation Unit that is
used to detect patterns or sequences on a given set of signals.
TGU is used to monitor a particular bus within a given region to
detect illegal transaction sequences or slave responses. It is also
used to monitor a data stream to detect protocol violations and to
provide a trigger point for centering data around a specific event
within the trace data buffer.

Signed-off-by: Mao Jinlong <quic_jinlmao@quicinc.com>
Signed-off-by: Hao Zhang <quic_hazha@quicinc.com>
Change-Id: I9e893689b51aa824aaf3063bb85b4e5d23808791